refactor(cargo_heather): split checker into submodules with shared strip helper
Break the 1142-line monolithic checker.rs into a checker/ module with
clear separation of concerns:

- matcher.rs: line-by-line header prefix matching
- extract.rs: pull the first comment block from file content (regular +
  script frontmatter), deduplicated via collect_comment_block
- strip.rs: remove existing headers via shared find_header_end helper;
  the previously parallel strip_existing_header / fix_script_content
  now share one algorithm with thin wrappers
- mod.rs: public API (check_files, check_file, fix_file, prepend_header,
  check_content) and dispatch helpers

The shared find_header_end uses take_while().count() instead of a
mutable += counter, so the mutation-test workaround comments are no
longer needed.

All 212 lib tests pass; clippy and rustdoc are clean.
